src/main/java/ru/yandex/practicum/filmorate/model/validation/FilmDurationValidator.java
package ru.yandex.practicum.filmorate.model.validation;

import javax.validation.ConstraintValidator;
import javax.validation.ConstraintValidatorContext;

public class FilmDurationValidator implements
        ConstraintValidator<PositiveDuration, Integer> {

    @Override
    public void initialize(PositiveDuration constraintAnnotation) {
        ConstraintValidator.super.initialize(constraintAnnotation);
    }

    @Override
    public boolean isValid(Integer duration, ConstraintValidatorContext constraintValidatorContext) {
        return duration >= 0;
    }
}
